Moving common elements from cores to abstract core.
