+++ src/Drupal/Driver/Cores/AbstractCore.php
@@ -7,9 +7,50 @@
 
 namespace Drupal\Driver\Cores;
 
+use Drupal\Component\Utility\Random;
 use Symfony\Component\DependencyInjection\Container;
 
 abstract class AbstractCore implements CoreInterface {
+
+  /**
+   * System path to the Drupal installation.
+   *
+   * @var string
+   */
+  private $drupalRoot;
+
+  /**
+   * URI for the Drupal installation.
+   *
+   * @var string
+   */
+  private $uri;
+
+  /**
+   * Random generator.
+   *
+   * @var \Drupal\Component\Utility\Random
+   */
+  private $random;
+
+  /**
+   * {@inheritDoc}
+   */
+  public function __construct($drupalRoot, $uri = 'default', Random $random = NULL) {
+    $this->drupalRoot = realpath($drupalRoot);
+    $this->uri = $uri;
+    if (!isset($random)) {
+      $random = new Random();
+    }
+    $this->random = $random;
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  public function getRandom() {
+    return $this->random;
+  }
 
   /**
    * {@inheritDoc}
